Start Project and Begin Header
The project is started with the index and style files necessary.
I started with the header and appropriately sized, colored, and fonted
the logo. I need to fix the layout with the links and I need to ensure
that the div takes up the whole page. I believe I should check
the * {} properties in previous excercises.