Add hero content begin hero styling
Added the content for the hero section and began the styling starting
with the flex declaration and the background color. There needs
to be work done withtthe image and styling the text should be done next.
Note that there is an h1 tag in both the hero and the header. Make sure
that h1 properties are used appropriately if they are used.