Add Schedulin Wizard specific text and styling
Began converting the web page from a general page to a business
specific one. Began with the logo and the header links then the
hero text. Need to add an image and a description of the company,
then add the co-founders pictures with breif discriptions, then add
a quote from Ivanna.